requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
Jinja2==3.1.1
jsonschema==4.4.0
pymongo==4.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
